requirements.txt.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.23.5
pyautogui>=0.9.53
pillow>=9.5.0
matplotlib>=3.7.1
pyyaml>=6.0
tqdm>=4.65.0
keyboard>=0.13.5
mss>=6.1.0